Widget Types: move useWidgetTypes into hooks/
