+++ routes/dashboard/widget-types/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useWidgetTypes } from './use-widget-types';
