+++ routes/dashboard/widget-types/hooks/use-widget-types.ts
@@ -6,8 +6,8 @@ import { useSelect } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import { store } from './store';
-import type { WidgetType } from './types';
+import { store } from '../store';
+import type { WidgetType } from '../types';
 
 export function useWidgetTypes(): WidgetType[] {
 	return useSelect( ( select ) => select( store ).getWidgetTypes(), [] );
